Moves TSDB to the stores package.
