+++ pkg/storage/stores/tsdb/pool.go
@@ -3,7 +3,7 @@ package tsdb
 import (
 	"sync"
 
-	"github.com/grafana/loki/pkg/storage/tsdb/index"
+	"github.com/grafana/loki/pkg/storage/stores/tsdb/index"
 )
 
 var (
