+++ pkg/storage/stores/tsdb/querier.go
@@ -22,7 +22,7 @@ import (
 	"github.com/prometheus/prometheus/model/labels"
 	"github.com/prometheus/prometheus/storage"
 
-	"github.com/grafana/loki/pkg/storage/tsdb/index"
+	"github.com/grafana/loki/pkg/storage/stores/tsdb/index"
 )
 
 // Bitmap used by func isRegexMetaCharacter to check whether a character needs to be escaped.
