+++ pkg/storage/stores/tsdb/testutil/objstore.go
